@@ -10,66 +10,64 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "prepare": "rmdir-cli src/assets/designer && ncp node_modules/@rocketshow/designer/assets src/assets/designer",
-    "prod": "node node_modules/@angular/cli/bin/ng build --prod"
+    "prod": "node node_modules/@angular/cli/bin/ng build --configuration production"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.13",
-    "@angular/common": "^9.1.13",
-    "@angular/compiler": "^9.1.13",
-    "@angular/core": "^9.1.13",
-    "@angular/forms": "^9.1.13",
-    "@angular/platform-browser": "^9.1.13",
-    "@angular/platform-browser-dynamic": "^9.1.13",
-    "@angular/router": "^9.1.13",
-    "@ngx-translate/core": "^10.0.2",
+    "@angular/animations": "^17.3.11",
+    "@angular/common": "^17.3.11",
+    "@angular/compiler": "^17.3.11",
+    "@angular/core": "^17.3.11",
+    "@angular/forms": "^17.3.11",
+    "@angular/platform-browser": "^17.3.11",
+    "@angular/platform-browser-dynamic": "^17.3.11",
+    "@angular/router": "^17.3.11",
+    "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^3.0.1",
-    "@rocketshow/designer": "^1.32.0",
+    "@rocketshow/designer": "^1.33.0",
     "angular2-websocket": "^0.9.7",
     "bootstrap": "^4.1.2",
     "core-js": "^2.5.7",
     "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
     "jquery": "^3.5.1",
-    "ngx-bootstrap": "^6.2.0",
-    "ngx-bootstrap-slider": "^1.9.0",
-    "ngx-dropzone-wrapper": "^10.0.1",
-    "ngx-sortablejs": "^11.1.0",
-    "ngx-toastr": "^8.10.2",
+    "ngx-bootstrap": "^12.0.0",
+    "ngx-bootstrap-slider": "^2.1.0",
+    "ngx-dropzone-wrapper": "^16.0.0",
+    "ngx-sortablejs-plus": "^15.0.2",
+    "ngx-toastr": "^19.0.0",
     "ngx-translate-multi-http-loader": "^3.0.0",
     "rxjs": "^6.6.3",
     "sortablejs": "^1.9.0",
     "split.js": "^1.5.10",
-    "three": "^0.104.0",
-    "tslib": "^1.10.0",
+    "three": "^0.165.0",
+    "tslib": "^2.0.0",
     "wavesurfer.js": "^3.0.0",
-    "web-animations-js": "^2.3.2",
     "web-audio-beat-detector": "^8.1.1",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.14",
-    "@angular/cli": "^9.1.14",
-    "@angular/compiler-cli": "^9.1.13",
-    "@angular/language-service": "^9.1.13",
-    "@types/jasmine": "~2.5.53",
+    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular/cli": "^17.3.8",
+    "@angular/compiler-cli": "^17.3.11",
+    "@angular/language-service": "^17.3.11",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.3",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "~2.1.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "^6.4.3",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "ncp": "^2.0.0",
-    "node-sass": "^4.9.2",
-    "protractor": "~5.3.0",
+    "protractor": "~7.0.0",
     "rmdir-cli": "^2.0.6",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "3.6.4"
+    "tslint": "~6.1.3",
+    "typescript": "5.4.5"
   }
 }
