@@ -2,19 +2,14 @@ package com.ascargon.rocketshow.api;
 
 import com.ascargon.rocketshow.composition.LeadSheet;
 import com.ascargon.rocketshow.composition.LeadSheetService;
-import org.apache.commons.fileupload.FileItemIterator;
-import org.apache.commons.fileupload.FileItemStream;
-import org.apache.commons.fileupload.servlet.ServletFileUpload;
-import org.apache.commons.io.FilenameUtils;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.FileInputStream;
-import java.io.InputStream;
 import java.util.List;
 
 @RestController()
@@ -22,10 +17,17 @@ import java.util.List;
 @CrossOrigin
 public class LeadSheetController {
 
+    private final ControllerService controllerService;
     private final LeadSheetService leadSheetService;
 
-    public LeadSheetController(LeadSheetService leadSheetService) {
+    public LeadSheetController(ControllerService controllerService, LeadSheetService leadSheetService) {
+        this.controllerService = controllerService;
         this.leadSheetService = leadSheetService;
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleException(Exception exception) {
+        return controllerService.handleException(exception);
     }
 
     @GetMapping("list")
@@ -34,26 +36,21 @@ public class LeadSheetController {
     }
 
     @PostMapping("upload")
-    public LeadSheet upload(HttpServletRequest request) throws Exception {
-        ServletFileUpload upload = new ServletFileUpload();
-        FileItemIterator itemIterator = upload.getItemIterator(request);
-
-        while (itemIterator.hasNext()) {
-            FileItemStream item = itemIterator.next();
-            String fileName = item.getName();
-
-            if (fileName != null) {
-                fileName = FilenameUtils.getName(fileName);
-            }
-
-            InputStream stream = item.openStream();
-
-            if (!item.isFormField()) {
-                return leadSheetService.saveLeadSheet(stream, fileName);
-            }
+    public LeadSheet upload(
+            @RequestParam("file") MultipartFile file,
+            @RequestParam("dzchunkindex") Long dzchunkindex,
+            @RequestParam("dztotalchunkcount") Long dztotalchunkcount
+    ) throws Exception {
+        String fileName = file.getOriginalFilename();
+        LeadSheet leadSheet = null;
+        if (dzchunkindex == 0) {
+            leadSheetService.saveLeadSheetInit(fileName);
         }
-
-        return null;
+        leadSheetService.saveLeadSheetAddChunk(file.getInputStream(), fileName);
+        if (dzchunkindex.equals(dztotalchunkcount - 1)) {
+            leadSheet = leadSheetService.saveLeadSheetFinish(fileName);
+        }
+        return leadSheet;
     }
 
     @PostMapping("delete")
