@@ -11,11 +11,14 @@ import com.ascargon.rocketshow.midi.MidiDeviceInService;
 import com.ascargon.rocketshow.midi.MidiDeviceOutService;
 import com.ascargon.rocketshow.util.*;
 import jakarta.xml.bind.JAXBException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.FileInputStream;
 
@@ -24,6 +27,9 @@ import java.io.FileInputStream;
 @CrossOrigin
 class SystemController {
 
+    private final static Logger logger = LoggerFactory.getLogger(SystemController.class);
+
+    private final ControllerService controllerService;
     private final StateService stateService;
     private final SetService setService;
     private final PlayerService playerService;
@@ -40,8 +46,10 @@ class SystemController {
     private final SessionService sessionService;
     private final CompositionService compositionService;
     private final DesignerService designerService;
+    private final BackupService backupService;
 
-    public SystemController(StateService stateService, SetService setService, PlayerService playerService, RebootService rebootService, ShutdownService shutdownService, SettingsService settingsService, MidiDeviceInService midiDeviceInService, MidiDeviceOutService midiDeviceOutService, UpdateService updateService, FactoryResetService factoryResetService, LogDownloadService logDownloadService, DiskSpaceService diskSpaceService, OperatingSystemInformationService operatingSystemInformationService, SessionService sessionService, CompositionService compositionService, DesignerService designerService) {
+    public SystemController(ControllerService controllerService, StateService stateService, SetService setService, PlayerService playerService, RebootService rebootService, ShutdownService shutdownService, SettingsService settingsService, MidiDeviceInService midiDeviceInService, MidiDeviceOutService midiDeviceOutService, UpdateService updateService, FactoryResetService factoryResetService, LogDownloadService logDownloadService, DiskSpaceService diskSpaceService, OperatingSystemInformationService operatingSystemInformationService, SessionService sessionService, CompositionService compositionService, DesignerService designerService, BackupService backupService) {
+        this.controllerService = controllerService;
         this.stateService = stateService;
         this.setService = setService;
         this.playerService = playerService;
@@ -58,6 +66,12 @@ class SystemController {
         this.sessionService = sessionService;
         this.compositionService = compositionService;
         this.designerService = designerService;
+        this.backupService = backupService;
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleException(Exception exception) {
+        return controllerService.handleException(exception);
     }
 
     @PostMapping("reboot")
@@ -137,10 +151,7 @@ class SystemController {
     @GetMapping("download-logs")
     public ResponseEntity<Resource> downloadLogs() throws Exception {
         InputStreamResource resource = new InputStreamResource(new FileInputStream(logDownloadService.getLogsFile()));
-
-        return ResponseEntity
-                .ok()
-                .body(resource);
+        return ResponseEntity.ok().body(resource);
     }
 
     @GetMapping("disk-space")
@@ -151,6 +162,28 @@ class SystemController {
     @GetMapping("operating-system-information")
     public OperatingSystemInformation getOperatingSystemInformation() {
         return operatingSystemInformationService.getOperatingSystemInformation();
+    }
+
+    @GetMapping("create-backup")
+    public ResponseEntity<Resource> createBackup() throws Exception {
+        InputStreamResource resource = new InputStreamResource(new FileInputStream(backupService.create()));
+        return ResponseEntity.ok().body(resource);
+    }
+
+    @PostMapping("restore-backup")
+    public ResponseEntity<Void> restoreBackup(
+            @RequestParam("file") MultipartFile file,
+            @RequestParam("dzchunkindex") Long dzchunkindex,
+            @RequestParam("dztotalchunkcount") Long dztotalchunkcount
+    ) throws Exception {
+        if (dzchunkindex == 0) {
+            backupService.restoreInit();
+        }
+        backupService.restoreAddChunk(file.getInputStream());
+        if (dzchunkindex.equals(dztotalchunkcount - 1)) {
+            backupService.restoreFinish();
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
 }
