@@ -18,19 +18,8 @@
 * BROWSER POLYFILLS
 */
 
-
-/** IE10 and IE11 requires the following for NgClass support on SVG elements */
-// import 'classlist.js';  // Run `npm install --save classlist.js`.
-
 /** Evergreen browsers require these. **/
-import 'core-js/es6/reflect';
-
-
-/**
- * Required to support Web Animations `@angular/animation`.
- * Needed for: All but Chrome, Firefox and Opera. http://caniuse.com/#feat=web-animation
- **/
-import 'web-animations-js';  // Run `npm install --save web-animations-js`.
+import 'core-js/es6/reflect';  // Run `npm install --save web-animations-js`.
 
 
 
@@ -43,7 +32,7 @@ import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 // See https://stackoverflow.com/questions/54416133/is-it-possible-to-use-ngzone-noop-for-particular-function-component-in-an
 import './ngzone-flags';
 
-import 'zone.js/dist/zone';  // Included with Angular CLI.
+import 'zone.js';  // Included with Angular CLI.
 
 
 
